@@ -5,9 +5,9 @@
   "repository": "github:razdvapoka/tanya-portfolio",
   "version": "0.0.1",
   "scripts": {
-    "dev": "node server.js",
+    "dev": "next",
     "build": "next build",
-    "start": "NODE_ENV=production node server.js",
+    "start": "next start",
     "lint": "eslint .",
     "upgrade-interactive": "npm-check --update"
   },
@@ -17,12 +17,11 @@
     "animejs": "github:razdvapoka/anime#master",
     "defaults.css": "^1.2.0",
     "emotion-theming": "^10.0.7",
-    "express": "^4.16.4",
     "fontfaceobserver": "^2.0.13",
     "hoist-non-react-statics": "^3.3.0",
     "isomorphic-unfetch": "^3.0.0",
     "keycode-js": "^1.0.4",
-    "next": "^7.0.3",
+    "next": "^9.4.4",
     "pss": "^4.3.0",
     "pss-components": "^4.3.0",
     "react": "^16.8.4",
@@ -40,8 +39,7 @@
     "@babel/register": "^7.0.0",
     "babel-eslint": "^10.0.1",
     "babel-plugin-inline-import": "^3.0.0",
-    "contentful": "^7.4.3",
-    "dotenv": "^6.2.0",
+    "contentful": "^7.14.4",
     "eslint": "^5.15.2",
     "eslint-config-standard": "^12.0.0",
     "eslint-config-standard-jsx": "^6.0.2",
