@@ -1,23 +1,21 @@
-import { Box } from 'pss-components'
-import React from 'react'
-import getConfig from 'next/config'
-import fetch from 'isomorphic-unfetch'
+import { Box } from "pss-components";
+import React from "react";
+import getConfig from "next/config";
+import fetch from "isomorphic-unfetch";
 
-import { Header, Section } from '../components'
-
-const { publicRuntimeConfig } = getConfig()
+import { Header, Section } from "../components";
+import { getPage } from "../content-api.js";
 
 class About extends React.Component {
-  static async getInitialProps () {
-    const res = await fetch(publicRuntimeConfig.siteUrl + '/api/page/about')
-    const about = await res.json()
-    return { about }
+  static async getInitialProps() {
+    const about = await getPage("about");
+    return { about };
   }
 
-  render () {
-    const { about } = this.props
+  render() {
+    const { about } = this.props;
     return (
-      <Box postion='relative' mgt={1}>
+      <Box postion="relative" mgt={1}>
         <Box pdx={6}>
           <Header isHome={false} />
         </Box>
@@ -27,8 +25,8 @@ class About extends React.Component {
           ))}
         </main>
       </Box>
-    )
+    );
   }
 }
 
-export default About
+export default About;
